requirements.txt
tensorflow
pandas
numpy
pillow
flask
requests
waitress
mlflow
prefect